+++ src/pages/projects/ProjectList.tsx
@@ -2,6 +2,7 @@ import { useEffect, useState } from 'react';
 
 import type { Project } from '@customTypes/project';
 import { requiredJwtTokeninstance } from '@libs/axios/axios';
+import { getProjectList } from '@services/project';
 import { AxiosResponse } from 'axios';
 
 interface AxiosRes<ResponseType> {
@@ -14,13 +15,6 @@ const ProjectList = () => {
   const [projectList, setProjectList] = useState<Project[] | null>(null);
 
   useEffect(() => {
-    const getProjectList = async () => {
-      const response: AxiosResponse<
-        AxiosRes<Project[]>,
-        any
-      > = await requiredJwtTokeninstance.get('/api/user/project/get');
-      return response.data.value;
-    };
     getProjectList().then((data) => {
       if (data) setProjectList(data);
     });
