+++ src/services/project/index.tsx
@@ -1,0 +1,3 @@
+import getProjectList from './getProjectList';
+
+export { getProjectList };
